requirements.txt
# Gmail to Podcast - Full Requirements
# Includes all dependencies for email processing, podcast generation, and video creation

# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0

# Gmail API and authentication
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0

# AI/LLM APIs
google-generativeai>=0.8.0

# Audio processing
pydub>=0.25.1

# Video generation (optional but included in full install)
moviepy>=1.0.0,<2.0.0
pillow>=9.0.0

# Utilities
requests>=2.28.0
urllib3>=1.26.0,<3.0.0